+++ internal-spring-boot-starter/src/main/java/eu/fraho/spring/securityJwt/starter/SecurityJwtInternalAutoConfiguration.java
@@ -1,0 +1,26 @@
+/*
+ * MIT Licence
+ * Copyright (c) 2017 Simon Frankenberger
+ *
+ * Please see LICENCE.md for complete licence text.
+ */
+package eu.fraho.spring.securityJwt.starter;
+
+import eu.fraho.spring.securityJwt.config.JwtRefreshConfiguration;
+import eu.fraho.spring.securityJwt.service.InternalTokenStore;
+import eu.fraho.spring.securityJwt.service.RefreshTokenStore;
+import org.springframework.boot.autoconfigure.AutoConfigureAfter;
+import org.springframework.boot.autoconfigure.AutoConfigureBefore;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@AutoConfigureAfter(SecurityJwtBaseAutoConfiguration.class)
+@AutoConfigureBefore(SecurityJwtNoRefreshStoreAutoConfiguration.class)
+public class SecurityJwtInternalAutoConfiguration {
+    @Bean
+    public RefreshTokenStore internalTokenStore(final JwtRefreshConfiguration jwtRefreshConfiguration) {
+        return new InternalTokenStore(jwtRefreshConfiguration);
+    }
+}